Change return value of `InsertMany` from `int64` to more conventional `int` (breaking)
Here, change the type of `Client.InsertMany` functions from `int64`s to
a more conventional `int`. This isn't really super necessary, but the
`int64` was an artifact of an internal value returned by sqlc, and using
`int` instead (which is also an int64 on the vast majority of all
system) is more conventional.

Because callers almost always know the size of the slice they passed to
`InsertMany`, or can ascertain it easily, I expect this change to break
very few integrations in practice. As we can see from our own test
suite, little had to change because `InsertMany`'s return value is
rarely needed.

+++ client.go
@@ -1342,7 +1342,7 @@ type InsertManyParams struct {
 //	if err != nil {
 //		// handle error
 //	}
-func (c *Client[TTx]) InsertMany(ctx context.Context, params []InsertManyParams) (int64, error) {
+func (c *Client[TTx]) InsertMany(ctx context.Context, params []InsertManyParams) (int, error) {
 	if !c.driver.HasPool() {
 		return 0, errNoDriverDBPool
 	}
@@ -1374,7 +1374,7 @@ func (c *Client[TTx]) InsertMany(ctx context.Context, params []InsertManyParams)
 // This variant lets a caller insert jobs atomically alongside other database
 // changes. An inserted job isn't visible to be worked until the transaction
 // commits, and if the transaction rolls back, so too is the inserted job.
-func (c *Client[TTx]) InsertManyTx(ctx context.Context, tx TTx, params []InsertManyParams) (int64, error) {
+func (c *Client[TTx]) InsertManyTx(ctx context.Context, tx TTx, params []InsertManyParams) (int, error) {
 	insertParams, err := c.insertManyParams(params)
 	if err != nil {
 		return 0, err
